requirements.txt
spotipy
python-dotenv
rich